split `calculate_rdf` and `calculate_all_pair_rdfs` into separate calc-rdf.ts file
